requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
cffi==1.17.1
    # via
    #   -r requirements.in
    #   cryptography
    #   pynacl
cryptography==45.0.6
    # via -r requirements.in
pycparser==2.22
    # via
    #   -r requirements.in
    #   cffi
pynacl==1.5.0
    # via -r requirements.in
